Reimplement the name demangling flag
