+++ src/ccc/stabs_to_ast.h
@@ -5,7 +5,6 @@
 
 #include "ast.h"
 #include "stabs.h"
-#include "mdebug_symbols.h"
 
 namespace ccc {
