+++ src/ccc/symbol_table.cpp
@@ -50,77 +50,6 @@ const SymbolTableFormatInfo* symbol_table_format_from_section(const char* sectio
 
 // *****************************************************************************
 
-const std::vector<ImporterFlagInfo> IMPORTER_FLAGS = {
-	{DONT_DEDUPLICATE_SYMBOLS, "--dont-deduplicate-symbols", {
-		"Do not deduplicate matching symbols from",
-		"different symbol tables."
-	}},
-	{DONT_DEDUPLICATE_TYPES, "--dont-deduplicate-types", {
-		"Do not deduplicate types from different",
-		"translation units."
-	}},
-	{DONT_DEMANGLE_NAMES, "--dont-demangle-names", {
-		"Do not demangle function names, global variable",
-		"names, or overloaded operator names."
-	}},
-	{INCLUDE_GENERATED_MEMBER_FUNCTIONS, "--include-generated-functions", {
-		"Output member functions that were likely",
-		"automatically generated by the compiler."
-	}},
-	{NO_ACCESS_SPECIFIERS, "--no-access-specifiers", {
-		"Do not print access specifiers."
-	}},
-	{NO_MEMBER_FUNCTIONS, "-no-member-functions", {
-		"Do not print member functions."
-	}},
-	{STRICT_PARSING, "--strict", {
-		"Enable strict parsing, which makes certain types",
-		"of errors that are likely to be caused by",
-		"compiler bugs fatal."
-	}},
-	{TYPEDEF_ALL_ENUMS, "--typedef-all-enums", {
-		"Force all emitted C++ enums to be defined using",
-		"a typedef. With STABS, it is not always possible",
-		"to determine if an enum was like this in the",
-		"original source code, so this option should be",
-		"useful for reverse engineering C projects."
-	}},
-	{TYPEDEF_ALL_STRUCTS, "--typedef-all-structs", {
-		"Force all emitted C++ structure types to be",
-		"defined using a typedef."
-	}},
-	{TYPEDEF_ALL_UNIONS, "--typedef-all-unions", {
-		"Force all emitted C++ union types to be defined",
-		"using a typedef."
-	}}
-};
-
-u32 parse_importer_flag(const char* argument)
-{
-	for(const ImporterFlagInfo& flag : IMPORTER_FLAGS) {
-		if(strcmp(flag.argument, argument) == 0) {
-			return flag.flag;
-		}
-	}
-	return NO_IMPORTER_FLAGS;
-}
-
-void print_importer_flags_help(FILE* out)
-{
-	for(const ImporterFlagInfo& flag : IMPORTER_FLAGS) {
-		fprintf(out, "\n");
-		fprintf(out, "  %-29s ", flag.argument);
-		for(size_t i = 0; i < flag.help_text.size(); i++) {
-			if(i > 0) {
-				fprintf(out, "                                ");
-			}
-			fprintf(out, "%s\n", flag.help_text[i]);
-		}
-	}
-}
-
-// *****************************************************************************
-
 Result<std::unique_ptr<SymbolTable>> create_elf_symbol_table(
 	const ElfSection& section, const ElfFile& elf, SymbolTableFormat format)
 {
@@ -209,7 +138,7 @@ Result<void> MdebugSymbolTable::import_symbol_table(
 	DemanglerFunctions demangler) const
 {
 	return mdebug::import_symbol_table(
-		database, m_image, m_section_offset, source, importer_flags, demangler);
+		database, m_image, m_section_offset, source, importer_flags | DONT_DEDUPLICATE_SYMBOLS, demangler);
 }
 
 Result<void> MdebugSymbolTable::print_headers(FILE* out) const
@@ -290,7 +219,7 @@ Result<void> SNDLLSymbolTable::import_symbol_table(
 	const u32 importer_flags,
 	DemanglerFunctions demangler) const
 {
-	return import_sndll_symbols(database, *m_sndll, source, demangler);
+	return import_sndll_symbols(database, *m_sndll, source, importer_flags, demangler);
 }
 
 Result<void> SNDLLSymbolTable::print_headers(FILE* out) const
